backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.49
pydantic-settings==2.9.1
python-multipart==0.0.20
httpx==0.28.1
PyJWT[crypto]==2.10.1
yfinance==0.2.58
pg8000==1.31.2
google-cloud-storage==3.1.0
